requirements.txt
defusedxml==0.6.0
mypy==0.790
mypy-extensions==0.4.3
python-dateutil==2.8.1
six==1.15.0
typed-ast==1.4.1
typing-extensions==3.7.4.3
